translate to Swift part 3
