--- Wire-iOS/Sources/UserInterface/ConversationList/ListContent/ConversationListContentController/Model/ConversationListViewModel+ZMConversationListObserver.swift
@@ -1,47 +1,0 @@
-
-// Wire
-// Copyright (C) 2019 Wire Swiss GmbH
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see http://www.gnu.org/licenses/.
-//
-
-import Foundation
-
-fileprivate let log = ZMSLog(tag: "ConversationListViewModel")
-
-extension ConversationListViewModel: ZMConversationListObserver {
-
-    public func conversationInsideList(_ list: ZMConversationList, didChange changeInfo: ConversationChangeInfo) {
-        delegate?.listViewModel(self, didUpdateConversationWithChange: changeInfo)
-    }
-
-    public func conversationListDidChange(_ changeInfo: ConversationListChangeInfo) {
-        guard let userSession = ZMUserSession.shared() else { return }
-
-        ///TODO: check the info instead of compare the lists
-        if changeInfo.conversationList == ZMConversationList.conversations(inUserSession: userSession) {
-            // If the section was empty in certain cases collection view breaks down on the big amount of conversations,
-            // so we prefer to do the simple reload instead.
-
-            updateConversationListAnimated() ///TODO: update oneOneOne also
-        } else if changeInfo.conversationList == ZMConversationList.pendingConnectionConversations(inUserSession: userSession) {
-            log.info("RELOAD contact requests")
-
-            let sectionIndex = SectionIndex.contactRequests
-            updateSection(sectionIndex)
-
-            delegate?.listViewModel(self, didUpdateSectionForReload: sectionIndex.sectionNumber(isFolderEnable: isFolderEnable))
-        }
-    }
-}
