+++ Wire-iOS/Sources/UserInterface/ConversationList/ListContent/ConversationListContentController/Model/ConversationListViewModel.swift
@@ -18,8 +18,42 @@
 
 import Foundation
 
-extension ConversationListViewModel {
-    @objc
+// Placeholder for conversation requests item
+final class ConversationListConnectRequestsItem : NSObject {}
+
+final class ConversationListViewModel: NSObject {
+    let contactRequestsItem: ConversationListConnectRequestsItem = ConversationListConnectRequestsItem()
+    /// ZMConversaton or ConversationListConnectRequestsItem
+    private(set) var selectedItem: Any?
+    weak var delegate: ConversationListViewModelDelegate?
+
+    private weak var selfUserObserver: NSObjectProtocol?
+    private var isFolderEnable = false
+    private var aggregatedItems: AggregateArray!
+    // Local copies of the lists.
+
+    ///TODO: [[Any]]
+    ///TODO: non optional
+    private var inbox: [Any]? = []
+    private var conversations: [Any]? = []
+    private var oneOnOneConversations: [Any]? = []
+
+    private var pendingConversationListObserverToken: Any?
+    private var conversationListObserverToken: Any?
+    private var clearedConversationListObserverToken: Any?
+    private var conversationListsReloadObserverToken: Any?
+
+
+    override init() {
+        super.init()
+        isFolderEnable = true
+
+        updateAllSections()
+        setupObserversForListReloading()
+        setupObserversForActiveTeam()
+        subscribeToTeamsUpdates()
+    }
+
     func setupObserversForListReloading() {
         guard let userSession = ZMUserSession.shared() else {
             return
@@ -28,7 +62,6 @@ extension ConversationListViewModel {
         conversationListsReloadObserverToken = ConversationListChangeInfo.add(observer: self, userSession: userSession)
     }
 
-    @objc
     func setupObserversForActiveTeam() {
         guard let userSession = ZMUserSession.shared() else {
             return
@@ -41,7 +74,7 @@ extension ConversationListViewModel {
         clearedConversationListObserverToken = ConversationListChangeInfo.add(observer: self, for: ZMConversationList.clearedConversations(inUserSession: userSession), userSession: userSession)
     }
 
-    func sectionCount() -> UInt {
+    var sectionCount: UInt {
         return aggregatedItems.numberOfSections()
     }
 
@@ -51,7 +84,7 @@ extension ConversationListViewModel {
 
     @objc(sectionAtIndex:)
     func section(at sectionIndex: UInt) -> [Any]? {
-        if sectionIndex >= sectionCount() {
+        if sectionIndex >= sectionCount {
             return nil
         }
         return aggregatedItems.section(at: sectionIndex)
@@ -84,14 +117,14 @@ extension ConversationListViewModel {
         return result
     }
 
-    func newConversationList() -> [ZMConversation]? {
-        guard let userSession = ZMUserSession.shared() else { return nil }
+    func newConversationList() -> [ZMConversation] {
+        guard let userSession = ZMUserSession.shared() else { return [] }
 
-        return ZMConversationList.conversations(inUserSession: userSession).map { $0 } as? [ZMConversation]
+        return ZMConversationList.conversations(inUserSession: userSession).map { $0 } as? [ZMConversation] ?? []
     }
 
-    func newOneOnOneConversationList() -> [ZMConversation]? {
-        return ZMUserSession.shared()?.oneOnOneConversations.map { $0 }
+    func newOneOnOneConversationList() -> [ZMConversation] {
+        return ZMUserSession.shared()?.oneOnOneConversations.map { $0 } ?? []
     }
 
     func reload() {
@@ -99,10 +132,6 @@ extension ConversationListViewModel {
         setupObserversForActiveTeam()
         //        debugLog("RELOAD conversation list")
         delegate?.listViewModelShouldBeReloaded()
-    }
-
-    func conversationListsDidReload() {
-        reload()
     }
 
     /// Select the item at an index path
@@ -200,4 +229,174 @@ extension ConversationListViewModel {
         }
     }
 
+    ///TODO: for debug
+    func toggleSort() {
+        isFolderEnable = !isFolderEnable
+
+        updateAllSections()
+        delegate?.listViewModelShouldBeReloaded()
+    }
+
+
+    @objc
+    func updateAllSections() {
+        updateSection(sectionIndexs: SectionIndex.allCases)
+    }
+
+    func updateSection(_ sectionIndex: SectionIndex, withItems items: [Any]? = nil) {
+        updateSection(sectionIndexs: [sectionIndex], withItems: items)
+    }
+
+    /// This updates a specific section in the model, by copying the contents locally.
+    /// Passing in a value of SectionIndexAll updates all sections. The reason why we need to keep
+    /// local copies of the lists is that we get separate notifications for each list,
+    /// which means that an update to one can render the collection view out of sync with the datasource.
+    ///
+    /// - Parameter sectionIndex: the section to update
+    func updateSection(sectionIndexs: [SectionIndex], withItems items: [Any]? = nil) {
+        if sectionIndexs == SectionIndex.allCases && items != nil {
+            assert(true, "Update for all sections with proposed items is not allowed.")
+        }
+
+        for sectionIndex in sectionIndexs {
+            switch sectionIndex {
+            case .contactRequests:
+                if let userSession = ZMUserSession.shared(), ZMConversationList.pendingConnectionConversations(inUserSession: userSession).count > 0 {
+                    inbox = items ?? [contactRequestsItem]
+                } else {
+                    inbox = []
+                }
+            case .conversations:
+                // Make a new copy of the conversation list
+                conversations = items ?? newConversationList()
+            case .contactsConversations:
+                oneOnOneConversations = items ?? newOneOnOneConversationList()
+            }
+        }
+
+
+        ///TODO: use a dictionary instead of fix size array
+        // Re-create the aggregate array
+        let sections: [Any]
+
+        if isFolderEnable {
+            sections = [inbox ?? [],
+                        oneOnOneConversations ?? [],
+                        conversations ?? []]
+        } else {
+            sections = [inbox ?? [],
+                        conversations ?? []]
+        }
+
+        aggregatedItems = AggregateArray(sections: sections)
+    }
+
+    @discardableResult
+    func updateForConvoType(sectionIndex: SectionIndex) -> Bool {
+        let sectionNumber = sectionIndex.sectionNumber(isFolderEnable: isFolderEnable)
+        let newList = newConversationList()
+
+        if let oldConversationList = aggregatedItems.section(at: sectionNumber) as? [ZMConversation],
+            oldConversationList != newList {
+
+            ///TODO: use diff kit and retire requiresReload
+            let startState = ZMOrderedSetState(orderedSet: NSOrderedSet(array: oldConversationList))
+            let endState = ZMOrderedSetState(orderedSet: NSOrderedSet(array: newList))
+            let updatedState = ZMOrderedSetState(orderedSet: [])
+
+            guard let changedIndexes = ZMChangedIndexes(start: startState, end: endState, updatedState: updatedState, moveType: ZMSetChangeMoveType.uiCollectionView) else { return true}
+
+            if changedIndexes.requiresReload == true {
+                reload()
+            } else {
+                // We need to capture the state of `newConversationList` to make sure that we are updating the value
+                // of the list to the exact new state.
+                // It is important to keep the data source of the collection view consistent, since
+                // any inconsistency in the delta update would make it throw an exception.
+                let modelUpdates = {
+                    self.updateSection(sectionIndex, withItems: newList)
+                }
+                delegate?.listViewModel(self, didUpdateSection: sectionNumber, usingBlock: modelUpdates, with: changedIndexes)
+            }
+
+            return true
+        }
+
+        return false
+    }
+
+    func updateConversationListAnimated() {
+        ///TODO: folder and non folder mode handling
+        if numberOfItems(inSection: SectionIndex.conversations.sectionNumber(isFolderEnable: isFolderEnable)) == 0 &&
+            numberOfItems(inSection: SectionIndex.contactsConversations.sectionNumber(isFolderEnable: isFolderEnable)) == 0 {
+
+            reload()
+        } else if !updateForConvoType(sectionIndex: .conversations) {
+            updateForConvoType(sectionIndex: .contactsConversations)
+        }
+    }
+
+    @objc(selectItem:)
+    @discardableResult
+    func select(itemToSelect: Any?) -> Bool {
+        guard let itemToSelect = itemToSelect else {
+            selectedItem = nil
+            delegate?.listViewModel(self, didSelectItem: nil)
+
+            return false
+        }
+
+        // Couldn't find the item
+        if self.indexPath(for: itemToSelect as! NSObject) == nil {
+            (itemToSelect as? ZMConversation)?.unarchive()
+        }
+
+        selectedItem = itemToSelect
+        delegate?.listViewModel(self, didSelectItem: itemToSelect)
+
+        return true
+    }
+
+    @objc
+    func subscribeToTeamsUpdates() {
+        if let session = ZMUserSession.shared() {
+            selfUserObserver = UserChangeInfo.add(observer: self, for: ZMUser.selfUser(), userSession: session)
+        }
+    }
+
+}
+
+fileprivate let log = ZMSLog(tag: "ConversationListViewModel")
+
+extension ConversationListViewModel: ZMConversationListObserver {
+
+    public func conversationInsideList(_ list: ZMConversationList, didChange changeInfo: ConversationChangeInfo) {
+        delegate?.listViewModel(self, didUpdateConversationWithChange: changeInfo)
+    }
+
+    public func conversationListDidChange(_ changeInfo: ConversationListChangeInfo) {
+        guard let userSession = ZMUserSession.shared() else { return }
+
+        ///TODO: check the info instead of compare the lists
+        if changeInfo.conversationList == ZMConversationList.conversations(inUserSession: userSession) {
+            // If the section was empty in certain cases collection view breaks down on the big amount of conversations,
+            // so we prefer to do the simple reload instead.
+
+            updateConversationListAnimated() ///TODO: update oneOneOne also
+        } else if changeInfo.conversationList == ZMConversationList.pendingConnectionConversations(inUserSession: userSession) {
+            log.info("RELOAD contact requests")
+
+            let sectionIndex = SectionIndex.contactRequests
+            updateSection(sectionIndex)
+
+            delegate?.listViewModel(self, didUpdateSectionForReload: sectionIndex.sectionNumber(isFolderEnable: isFolderEnable))
+        }
+    }
+}
+
+
+extension ConversationListViewModel: ZMConversationListReloadObserver {
+    func conversationListsDidReload() {
+        reload()
+    }
 }
